Tests: Removed a TC which does not make sense.
This TC is unnecessary. It accepts both types of selection (text and none), so both can be considered as correct, but
expects the type to be none on IEs and text on other browsers. Changing the selection behaviour made the test red,
but that only shows that something has changed, not that something is incorrect. Let's avoid such TCs.